Crate query_structure

Re-exports

pub use psl;

pub use filter::*;

Modules

arithmetic

Note: Only number operations are implemented at the moment.

ast

The AST data structure. It aims to faithfully represent the syntax of a Prisma Schema, with source span information.

filter

Filtering types to select records from the database

prelude

walkers

Convenient access to a datamodel as understood by ParserDatabase.

Structs

CompositeSelection

ConversionFailure

DefaultValue

Represents a default specified on a field.

FieldSelection

A selection of fields from a model.

Fields

InternalDataModel

ManyRecords

ModelProjection

Projection of a Model. A projection is a (sub)set of fields of a model. There can only ever be fields of one model contained in a particular ModelProjection

NativeTypeInstance

Represents an instance of a native type declared in the Prisma schema.

OrderByRelevance

OrderByScalar

OrderByScalarAggregation

OrderByToManyAggregation

QueryArguments

QueryArguments define various constraints queried data should fulfill:

Record

RelationSelection

SelectionResult

Represents a set of results.

SingleRecord

ValueGenerator

Zipper

Enums

CompositeFieldId

DateType

DefaultKind

Represents a default specified on a field.

DomainError

Field

FieldArity

An arity of a data model field.

NullsOrder

OrderBy

OrderByHop

Describes a hop over to a relation or composite for an orderBy statement.

PrismaValue

RelationLoadStrategy

ScalarFieldId

SelectedField

A selected field. Can be contained on a model or composite type.

SortAggregation

SortOrder

TypeIdentifier

ValueGeneratorFn

Functions

convert

decode_bytes

dml_default_kind

encode_bytes

parse_datetime

Parses an RFC 3339 and ISO 8601 date and time string such as 1996-12-19T16:39:57-08:00, then returns a new DateTime with a parsed FixedOffset.

stringify_datetime

Stringify a date to the following format 1999-05-01T00:00:00.000Z

Type Aliases

CompositeField

CompositeFieldRef

CompositeType

InternalEnum

InternalEnumValue

Model

PrismaListValue

PrismaValueResult

Relation

RelationField

RelationFieldRef

Result

ScalarField

ScalarFieldRef