Module query_structure::filter

Filtering types to select records from the database

The creation of the types should be done with ScalarCompare and RelationCompare. CompositeCompare.

Structs

CompositeFilter

JsonCondition

OneRelationIsNullFilter

Filter that is solely responsible for checking if a to-one related record is null. Todo there’s no good, obvious reason why this is a separate filter.

RelationFilter

ScalarFilter

Filtering with a scalar value. From a GraphQL point of view this is in the head of the query:

ScalarListFilter

Enums

AggregationFilter

CompositeCondition

ConditionListValue

ConditionValue

Filter

JsonFilterPath

JsonTargetType

QueryMode

RelationCondition

ScalarCondition

ScalarListCondition

ScalarProjection

Traits

CompositeCompare

Comparison methods for composite fields.

IntoFilter

JsonCompare

Comparison methods for json fields

RelationCompare

Comparison methods for relational fields.

ScalarCompare

Comparing methods for scalar fields.

ScalarListCompare

Comparison methods for scalar list fields.