Enum query_structure::filter::Filter

pub enum Filter {
    And(Vec<Filter>),
    Or(Vec<Filter>),
    Not(Vec<Filter>),
    Scalar(ScalarFilter),
    ScalarList(ScalarListFilter),
    OneRelationIsNull(OneRelationIsNullFilter),
    Relation(RelationFilter),
    Composite(CompositeFilter),
    BoolFilter(bool),
    Aggregation(AggregationFilter),
    Empty,
}

Variants

And(Vec<Filter>)

Or(Vec<Filter>)

Not(Vec<Filter>)

Scalar(ScalarFilter)

ScalarList(ScalarListFilter)

OneRelationIsNull(OneRelationIsNullFilter)

Relation(RelationFilter)

Composite(CompositeFilter)

BoolFilter(bool)

Aggregation(AggregationFilter)

Empty

Implementations

impl Filter

pub fn and(filters: Vec<Filter>) -> Self

pub fn or(filters: Vec<Filter>) -> Self

pub fn not(filters: Vec<Filter>) -> Self

pub fn empty() -> Self

pub fn size(&self) -> usize

Returns the size of the topmost filter elements (does not recursively compute the size).

pub fn should_batch(&self, chunk_size: usize) -> bool

pub fn can_batch(&self) -> bool

pub fn batched(self, chunk_size: usize) -> Vec<Filter>

pub fn set_mode(&mut self, mode: QueryMode)

pub fn count(condition: Filter) -> Self

pub fn average(condition: Filter) -> Self

pub fn sum(condition: Filter) -> Self

pub fn min(condition: Filter) -> Self

pub fn max(condition: Filter) -> Self

pub fn as_scalar(&self) -> Option<&ScalarFilter>

pub fn into_scalar(self) -> Option<ScalarFilter>

pub fn is_empty(&self) -> bool

pub fn scalars(&self) -> Vec<ScalarFieldRef>

pub fn unique_scalars(&self) -> Vec<ScalarFieldRef>

pub fn has_relations(&self) -> bool

Returns true if filter contains conditions on relation fields.