Struct query_structure::ModelProjection

pub struct ModelProjection { /* private fields */ }

Projection of a Model. A projection is a (sub)set of fields of a model. There can only ever be fields of one model contained in a particular ModelProjection

Implementations

impl ModelProjection

pub fn new(fields: Vec<Field>) -> Self

pub fn new_from_scalar(fields: Vec<ScalarFieldRef>) -> Self

pub fn names(&self) -> impl Iterator<Item = &str>

Returns all field names (NOT database level column names!) of contained fields.

pub fn db_names(&self) -> impl Iterator<Item = String> + '_

Returns all database (e.g. column or field) names of contained fields.

pub fn fields(&self) -> impl Iterator<Item = &Field>

Returns an iterator over all fields contained in this projection.

pub fn scalar_length(&self) -> usize

Returns the length of scalar fields contained in this projection, e.g. the actual number of SQL columns or document fields for this model projection.

pub fn scalar_fields(&self) -> impl Iterator<Item = ScalarFieldRef> + '_

Returns an iterator over all scalar fields represented by this model projection, in order. Resolves relation fields to all backing scalar fields, if existing.

pub fn map_db_name(&self, name: &str) -> Option<ScalarFieldRef>

pub fn type_identifiers_with_arities(&self) -> Vec<(TypeIdentifier, FieldArity)>