Struct query_structure::SelectionResult

pub struct SelectionResult {
    pub pairs: Vec<(SelectedField, PrismaValue)>,
}

Represents a set of results.

Fields
pairs: Vec<(SelectedField, PrismaValue)>
Implementations

impl SelectionResult

pub fn new<T>(pairs: Vec<(T, PrismaValue)>) -> Selfwhere T: Into<SelectedField>,

pub fn add<T>(&mut self, pair: (T, PrismaValue))where T: Into<SelectedField>,

pub fn get(&self, selection: &SelectedField) -> Option<&PrismaValue>

pub fn values(&self) -> impl Iterator<Item = PrismaValue> + '_

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn db_names(&self) -> impl Iterator<Item = &str> + '_

pub fn split_into( self, field_selections: &[FieldSelection] ) -> Vec<SelectionResult>

Consumes this SelectionResult and splits it into a set of SelectionResults based on the passed FieldSelections. Assumes that the transformation can be done.

pub fn as_scalar_fields(&self) -> Option<Vec<ScalarFieldRef>>

Checks if self only contains scalar field selections and if so, returns them all in a list. If any other selection is contained, returns None.

pub fn coerce_values(self) -> Result<Self>

Coerces contained values to best fit their type.

Scalar fields coerce values based on the TypeIdentifier.

Composite fields must be objects and contained fields must also follow the type coherence.