Type Alias query_structure::walkers::ScalarFieldWalker

pub type ScalarFieldWalker<'db> = Walker<'db, ScalarFieldId>;

A scalar field, as part of a model.

Aliased Type

struct ScalarFieldWalker<'db> {
    pub db: &'db ParserDatabase,
    pub id: ScalarFieldId,
}

Fields
db: &'db ParserDatabase

The parser database being traversed.

id: ScalarFieldId

The identifier of the focused element.

Implementations

impl<'db> Walker<'db, ScalarFieldId>

pub fn field_id(self) -> FieldId

The ID of the field node in the AST.

pub fn ast_field(self) -> &'db Field

The field node in the AST.

pub fn is_unique(self) -> bool

Is this field unique? This method will return true if:

The field has an @id or @unique attribute.

There is an @@id or @@unique on the model that contains only this field.

pub fn name(self) -> &'db str

The name of the field.

pub fn default_attribute(self) -> Option<&'db Attribute>

The @default() AST attribute on the field, if any.

pub fn database_name(self) -> &'db str

The final database name of the field. See crate docs for explanations on database names.

pub fn is_autoincrement(self) -> bool

Does the field have an @default(autoincrement()) attribute?

pub fn is_single_pk(self) -> bool

Does the field define a primary key by its own.

pub fn is_part_of_a_compound_pk(self) -> bool

Is the field part of a compound primary key.

pub fn is_ignored(self) -> bool

Is there an @ignore attribute on the field?

pub fn is_optional(self) -> bool

Is the field optional / nullable?

pub fn is_list(self) -> bool

Is the field a list

pub fn is_updated_at(self) -> bool

Is there an @updatedAt attribute on the field?

pub fn field_type_as_enum(self) -> Option<Walker<'db, EnumId>>

Is this field’s type an enum? If yes, walk the enum.

pub fn mapped_name(self) -> Option<&'db str>

The name in the @map(<name>) attribute.

pub fn model(self) -> Walker<'db, ModelId>

The model that contains the field.

pub fn raw_native_type( self ) -> Option<(&'db str, &'db str, &'db [String], Span)>

(attribute scope, native type name, arguments, span)

For example: @db.Text would translate to (“db”, “Text”, &[], )

pub fn is_unsupported(self) -> bool

Is the type of the field Unsupported("...")?

pub fn default_value(self) -> Option<DefaultValueWalker<'db>>

The @default() attribute of the field, if any.

pub fn scalar_field_type(self) -> ScalarFieldType

The type of the field.

pub fn scalar_type(self) -> Option<ScalarType>

The type of the field in case it is a scalar type (not an enum, not a composite type).

impl<'db, I> Walker<'db, I>

pub fn walk<J>(self, other: J) -> Walker<'db, J>

Traverse something else in the same schema.